Initialize all values in BugReprotData.
On ToT I found that if I issued a bug report quickly it would crash because the syslogs data was not ready, presumably because fetching that data was put on a separate thread.
I'm trying to confirm that this is the same issue as the cras repport in chromium-os:15253, but regardless this simple fix should be in the M12 branch.

BUG=chromium-os:15253
TEST=Report an issue immediately after logging in. Chrome should not crash.

Review URL: http://codereview.chromium.org/7016045