Convert chromeos::BrowserView context menu to MenuItemView
This is the context menu from right-click in the tab bar for Restore
Tab, New Tab, and Task Manager.  The existing implementation used
views::Menu2 and this CL converts it to views::MenuItemView.

There is a little mismatch because the ::BrowserView superclass to chromeos::BrowserView
implements ui::SimpleMenuModel::Delegate, which sort of assumes that
subclasses are using a menu implementation that calls this delegate.
MenuItemView uses MenuDelegate instead, so I implemented most of the
MenuDelegate calls as an adapter to the ui::SimpleMenuModel::Delegate
implementation in the ::BrowserView class.

BUG=chromium-os:13887
TEST=none

Review URL: http://codereview.chromium.org/6873078