Add utility for converting SkCanvas to CGContext
Skia on Mac uses Skia to render WebKit, and CG 
to render UI elements. The CG elements need a 
transcribed graphics context that preserves the 
canvas matrix, and the canvas clip. 

The BitLockerSkia utility class sets up a CGContext 
from the SkCanvas, locks the bitmap's bits, and 
releases the lock when the class goes out of scope. 

Each time the CGContext is retrieved, it is rebuilt. 
This permits the caller to modify the same canvas 
between calls. 

Outside of the unit test, there are no callers
to this utility for now.

BUG=79463
TEST=SkiaUtilsMacTest.*

Review URL: http://codereview.chromium.org/7031006