GTTF: Get rid of IsBrowserRunning. It is inherently racy and should not be used. Where needed, explicitly wait for the browser to shut down.
The problem with IsBrowserRunning is the raciness: the browser may be alive when you call IsBrowserRunning, and crash right after the call completes. There are also various ways of implementing IsBrowserRunning. We could do a waitpid on it (that was the previously used solution), but then if the browser did exit we'd be losing information about its exit code, whether it was killed by a signal, and so on (we'd have to plumb this information from base/process_util through IsBrowserRunning to UI test infrastructure; that'd be rather complicated and error prone). The solution the code was using just before this removal was to send a dummy IPC message to the browser. But then the browser might be shutting down and no longer receiving IPC messages, but the process itself could still be running.

So the simplest way to deal with the problem is to get the browser's exit code in UITest::TearDown. This allows us to detect crashes on shutdown in a non-racy way, which is very important. And then if the browser dies in the middle of a test, the next IPC call to it will fail, and the shutdown check in UITest::TearDown will also fail.

BUG=none
Review URL: http://codereview.chromium.org/7039044