C++ Readability CL for erikwright.
Ready Mode is a feature of Chrome Frame that allows it to be installed in a semi-active state. The first time the user accesses, using IE, a site that supports Chrome Frame, the Ready Mode infobar prompt is displayed in IE, offering to permanently enable or disable Chrome Frame.

RegistryReadyModeState implements a strategy for implementing the user's requested state change by invoking the Chrome Frame installer.
 
BUG=None
TEST=None
Review URL: http://codereview.chromium.org/6733015