Culls the list of timestamps to most recent 50 before saving it to the db.
It looks like sync deals with a structure called |AutofillEntry| which seems not to be used elsewhere in the code. The rest of the autofill code uses |FormField|.

A naive implementation of checking the number of timestamps on add or update and culling them would not work with sync. Reason being sync can be the one that is modifying the item. Once the timestamps have been culled sync needs to know the new timestamps so it can update the sync db. To update the sync db you need to hold a sync transaction which we dont hold when we are updating the autofill entries. Reason being sync transaction can theoretically block the UI thread and we dont want to hold it when we are updating the db.

So there are 3 possible solutions:
1. Once the list of timestamps have been updated send a notification to sync. But we purposefully disable receiving any notifications because we normally dont want to be notified of changes that we are making.
2. Sync queries back these items later and figures out the timestamps stored in disk. This is a 'degenerate' form of (1).
3. When creating an object of |AutofillEntry| struct, cull the time stamps. This is an in memory object. So sync creates this object and if it finds this object's timestamps have been culled then it updates sync db with it and if this is browser start we also puts this entry in a list which will eventually be used to update the web db(after browser start we will not update the web db until the next restart). This is what I have implemented now.

However because of the issue that AutofillEntry struct is not used in the autofill code this would mean that if the clients dont have sync enabled they will not get the age out feature. If this needs to be fixed the following approach can be taken in autofill code:

use autofill entry structure for committing.(I mean inside the function AutofillTable::AddFormFieldValueTime) But this has some perf implications because every time we update we will have to query the database for the current timestamps. Not sure if this is a big deal. (We could make it such that we read the timestamps from db only if the count is greater than 50. Autofill code already queries the count in code for a different reason. But once you have reached 50 then every time we update we will have to query the list of timestamps).

If the perf implication is a big deal we should consider storing the autofill entries in memory. They should not be big because pretty soon we are going to age out autofill as well.



BUG=
TEST=


Review URL: http://codereview.chromium.org/6903134