LBYL on GSettings
Fixes a regression from r85968.

On my Ubuntu 10.10 box, g_settings_new does not fail gracefully if the schema is not found: it calls an assert. This checks if the schema exists first. The docs for g_settings_list_schemas specify not to modify or release the return value.

BUG=83376
TEST=by hand and trybots


Review URL: http://codereview.chromium.org/6990008