Revert 86306 (failed compile on win builder) - Move OpenGL ES 2.0 conformance test support into main tree
This support breaks often because it is not built but the buildbots.
This CL adds it to the main tree so it will be built by the buildbots.

Not sure where to add it in all.gyp but I need to make sure it's built

TEST=opengl es 2.0 conformance tests still build
BUG=83382

R=apatrick@chromium.org

Review URL: http://codereview.chromium.org/7057033

TBR=gman@chromium.org
Review URL: http://codereview.chromium.org/7066003