Make WebGraphics3DCommandBufferImpl use GLES2Implemenation directly.
Note: I can do this 2 ways. I can either make it call it truly
direclty as in

  gles2_implementation_->BindBuffer(args)

or I can do as I have using CALL_GL_API in which case it's

  CALL_GL_API(BindBuffer)(args)

The only advantage to the second format is a macro changes
it back to the way it was. On the other hand the class is
called WebGraphics3DCommandBufferImpl so maybe I should
change it to the first style.

TEST=try-bots and ran chrome myself
BUG=none

Review URL: http://codereview.chromium.org/7056020