Split UI-specific bits off ExtensionFunction into a separate class.
This is a precursor to another patch I'm working on, which will add
ExtensionFunctions that can run on the IO thread. The webRequest API will use
this.

I've also done a bit of other cleanup:
- moved almost everything off of {Async,Sync}ExtensionFunction, up into the
base class, since none of it was specific to either of those.
- store Extension directly on ExtensionFunction, rather than needing to look it
up, since it is now refcounted.

BUG=no
TEST=no

Review URL: http://codereview.chromium.org/7073001