Stop using the boring score for judging we should update a thumbnail or not.
The absolute boring score is not suitable for judging a thumbnail is actually
boring or not. For instance, www.google.com is very succinct so the boring
score can be as high as 0.9, where 1 is the maximum. This score should only
be used for comparing two thumbnails taken from the same page to see which one
is more boring/interesting.

BUG=65936
TEST=with --enable-in-browser-thumbnailing --vmodule=thumbnail_generator=1, confirm that the thumbnail is not updated very often for www.google.com.

Review URL: http://codereview.chromium.org/7071008