Added hidden network support to wifi compliance tests.
Also modified the checking of service_order. There are some problems where if the service is not present, it will not show after calling GetServiceOrder even though it was specifically set.  The current method is to make sure the order of the common elements of what we set and what is retrieved from the device is in the same order.

TODO: We need to clear the network list before running the tests.  Otherwise, hidden networks suddenly become visible if they are remembered and this will throw off the hidden network test.

BUG=None
TEST=Run it with one of the hidden networks

Review URL: http://codereview.chromium.org/7078017