Integrate file systems with BrowsingDataRemover
The functionality desired isn't entirely clear.  This CL takes care of the bits that everyone agrees on: when a user deletes her browsing data "since the beginning of time" both persistent and temporary filesystems are removed entirely.  What we do for other time periods is still up in the air.

BUG=63700
TEST=unit_tests

Review URL: http://codereview.chromium.org/7087020