gtk: Makeshift multiprofile button
Since the multiprofile UI is being redesigned, I didn't spend time making this pretty. It's just a normal gtk button with the sync name on it, instead of that fancy button mac and windows have. The button is only shown when --multi-profiles is passed in and the user is signed in to sync, so that should be ok.

BUG=83169
TEST=Start chrome with --multi-profiles. No button visible. Sign in to sync. Button appears. Clicking it produces same menu as on mac/windows. Sign out, button disappears.


Review URL: http://codereview.chromium.org/7017016