Adding chrome://sessions webui for prototyping tab sync NTP stuff.
This adds the necessary plumbing for chrome://sessions webui page which shows a hierarchy of sessions and their windows and tabs as well as a "magic" list of suggested tabs. The magic list is currently just the last 10 tabs by timestamp for now - but can later be changed to prototype other heuristics.

BUG=none
TEST=manual

Review URL: http://codereview.chromium.org/6969016