Mac: Fix icons in download dom UI
Icons in the download DOM UI were very fuzzy.

This was a regression due to r76743.

The problem was that IconLoad was returning a multi-resolution image but callers weren't multi-resolution aware.

For example, the download DOM UI would encode the image to PNG but it would just pick the first bitmap which happened to be 16x16. This caused icons in the DOM UI to look fuzzy.

I think there are two fixes for this:
1. Change IconLoader so that callers no longer specify the icon size. Instead, callers must pick the right resolution when they get the icon loaded callback.
2. Have callers specify that they want all resolutions.

I went with 2 because always loading all icon sizes can be expensive.

BUG=84741
TEST=Ran and verified that the downloads UI looks good.


Review URL: http://codereview.chromium.org/6995068