Changed typed url sync to no longer modify typed/visit_count.
The sync code now relies on the underlying history code to keep typed/visit_count
up-to-date. This requires us to sync across the transition type for each visit.

Since we want to set the transition type appropriately for every visit, this CL
also changes AddPagesWithDetails() to no longer automatically create a visit with
the last_visit timestamp if the VisitSource is SYNCED. For convenience, it still
creates a visit for the other sources (browser history import).

BUG=84258
TEST=run sync integration/unit tests, run typed url test cases.

Review URL: http://codereview.chromium.org/7104088