Re-enable thread IO restriction override.
The Linux Views browser ui tests are failing. The test dies on an thread io restriction violation. I'm guessing (with Backer's help) that this may be due to the bot not having GL enabled, and continuing attempts to init the bindings fail (normally it would never get to us, hence why we thought the IO restriction was gone).

BUG=linux views fails on compositor startup due to thread io restriction
TEST=browser ui tests

Review URL: http://codereview.chromium.org/7104115