Remove incorrect comments from the audio config headers.
Currently RecommendSampleFrameCount just clamps the value passed in, so passing in 0 returns the min frame count (64) instead of an actual value recommended by the system.  As a note, if/when this behavior changes we'll also need to update the NaCl audio walkthrough doc which details the current behavior.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/7046115