Make the web content accessibility tree a descendant of the main window's accessibility tree again, which makes it possible to debug Chrome accessibility on Windows using tools like AccExplorer32 or AccProbe.
This change refactors the code that retrieves the Windows IAccessible
interface from a View that wraps a native HWND. The previous code was
brittle, using ViewProps and hardcoded View classname tests. This
change moves that logic directly to the classes that wrap native
views.

BUG=54220
BUG=85673
TEST=Updated AccessibilityWinBrowserTest.ContainsRendererAccessibilityTree

Review URL: http://codereview.chromium.org/6995126