Enforce RGB even on buggy drivers.
Also STENCIL and DEPTH.

This might not be a complete solution. It assumes
that if we ask for an RGB backbuffer or renderbuffer
and the driver gives us RGBA that we can still clear
that RGBA to 1,1,1,1. After that it sets the masks
so that alpha can never be rendered to.

Similarly for DEPTH and STENCILS.

It's possible we should just always ask for RGBA, at
least for Renderbuffers.

TEST=unit tests, ran OpenGL ES 2.0 conformance test and WebGL conformance tests.
BUG=85385

R=apatrick@chromium.org

Review URL: http://codereview.chromium.org/7099007