Automatically put all GL contexts associated with a particular GPU channel (and renderer process) in the same share group.
This is work towards allowing offscreen canvas contexts to be allocated before compositor view contexts are created. This is a problem because a canvas might be created before the window it will be displayed in exists.

This does not fix the bug on its own.

BUG=80703
Review URL: http://codereview.chromium.org/7104148