Removed the GetProductVersion function from webkit_glue and replace it with the BuildUserAgent function. The BuildUserAgent function has been deleted from user_agent.cc.  The implementation of this function in content\renderer_glue.cc calls the GetUserAgent API in ContentClient which is implemented by the embedder (Chrome).
Added implementations of the BuildUserAgent function for test_shell and DumpRenderTree. To build the user agent
string we need the webkit major and minor versions. Added getters for them in the webkit_glue namespace in the
user_agent.h/.cc files.

This helps reduce the implicit dependency of content on chrome.

BUG=82454
Review URL: http://codereview.chromium.org/7166004