Add custom NSView as a titlebar of a Panel so we can custom-paint it.
The XIB file change is adding a view and linking it with PanelTitlebarViewCocoa class. The view is not parented to the Panel window since XIB can't set a custom view to shadow the regular titlebar - so we hook it up in awakeFromNib for the panel.

Review URL: http://codereview.chromium.org/7108069