Move the video decoder and video layer to the new API/thunk system.
This changes the video decoder API to make GetConfigs be a resource member
function so it can be routed with the rest of the resource functions.

This patch also removes the font_list_dev API which was never implemented and
has been replaced with a function on the Font interface.

TEST=manual
BUG=none
Review URL: http://codereview.chromium.org/7193018