Fix a plugin process crash which occurs in the destructor of the NPObjectProxy object.
Here we send out a sync message NPObjectMsg_Release to the renderer to destroy the corresponding stub.
In the context of this outgoing sync call we get an incoming sync message to create a new plugin instance. 

The plugin in this context retrieves the window object and eventually invokes on it. The NPObjectProxy instance
is reused if it points to same underlying NPObject. This is reused based on the existence of the NPObjectProxy
routing id in a map maintained by the plugin channel.

In this case we end up reusing an NPObjectProxy instance which is being destroyed. This results in a crash
when the call sequence unwinds.

Fix is to remove the NPObjectProxy routing id from the channel map before sending out the NPObjectMsg_Release
sync message. 

Fixes bug http://code.google.com/p/chromium/issues/detail?id=86576

BUG=86576
TEST=manually as described in the bug.
Review URL: http://codereview.chromium.org/7201013