Fix NativeThemeWin with --enable-accelerated-drawing path.
Do a local copy of the bitmap pixels to a temporary bitmap whose pixels are created with |allocPixels()| and hence refcounted.

This is necessary because SkPicture defers drawing until some time after the |drawBitmap()| call and expects the ref-counting mechanism to hold on to the pixels for it.

BUG=none
TEST=manual using --enable-accelerated-drawing and --force-compositing-mode flags

Review URL: http://codereview.chromium.org/7189064