Merge 89600 - Treat ERR_CONNECTION_CLOSED as a succesful download.
Some servers will send a Content-Length which exceeds the number of bytes sent in the message body.

Although this violates the RFC spec, other major browsers permit it
and treat a download as complete in that case. Chrome did as well
prior to M12, and now some users are unable to access the item that
they downloaded.

BUG=85549
TEST=Go to server which advertises larger Content-Length than number of bytes sent, confirm that the download is no longer reported as interrupted.


Review URL: http://codereview.chromium.org/7203004

TBR=cbentzel@chromium.org
Review URL: http://codereview.chromium.org/7234012