Set up Chrome PFQ builder onto vm2-m2
NOTE:  This commit requires a master restart for builds
that use x86-generic-chrome-pre-flight-queue

Here we add the chrome pfq.  I had to modify the chromeos factory
in order to get the triggering to work similiarly to the pfq which
is necessary for the syncing (therefore requiring the restart).

I also went ahead and took the change to remove all the old_style code.
Leaving the option in __init__ for backwards compatibility to other
master.cfg's but will eventually remove.

BUG=chromium-os:16816
TEST=Doing it now.

Review URL: http://codereview.chromium.org/7239004