Linux: remove the forced keyring unlock on password store initialization. It is no longer required for our gnome keyring integration to work correctly.
This has the side effect of avoiding an unlock prompt unless a login form would
actually be filled in by data in the keyring, lessening phishing vulnerability.

In order to make this work correctly, we also have to be a little more careful
about when we commit to using the keyring instead of the built-in store. In
particular, because it can succeed without unlocking when there is no data to
return (which is good, and avoids unexpected unlock prompts), we must require
that we actually got some data before assuming that we can use the keyring.

Note that we don't need to do any sort of merging or fall back to the built-in
store in these cases, because we know it's empty - if it had any data in it, we
would have migrated it, and we wouldn't be in this undecided state to begin
with.

This subtlety is probably the cause of some old repeated-unlock-prompt bugs
that were masked by the unlock forcing. (Which was part of its point.)

BUG=85285
Review URL: http://codereview.chromium.org/7189001