Enable selection of a candidate by mouse-down on ChromeOS candidate window.
BUG=chromium-os:11423
TEST=1. enable Japanese IME
2. click a textarea (e.g. google search form) in web content area.
   THIS ISSUE DOES NOT REPRODUCE ON OMNIBOX.
3. type 'a' space space
4. click one of the candidates in the IME candidate window
5. Verify the clicked candidate is selected.

Before the embedding of the candidate window inside chrome's process
http://src.chromium.org/viewvc/chrome?view=rev&revision=72934
selection was done at the timing of mouse-UP. After the embedding,
the selection-by-mouse feature is completely broken. This patch is
a fix for this, by enabling the selection at mouse-DOWN time,
because it turned out difficult to keep the behavior (see the comments
in the code).

Review URL: http://codereview.chromium.org/7248001