Re-Fixes Texture::Copy
The issue is Texture::Copy is being called without calling the
required Texture::AllocateStorage. This happens to work because
glCopyTexImage2D also allocates storage.

Al would know the correct way to fix this so Texture::AllocateStorage
is always called but he's in the process of ripping out all this
code so this CL just makes it work as is (which is what it was
doing before except now it has the correct format)

https://bugs.webkit.org/show_bug.cgi?id=63638
https://bugs.webkit.org/show_bug.cgi?id=63643

TEST=ran DRT with layout tests mentioned in bugs
BUG=none (see above)


Review URL: http://codereview.chromium.org/7218056