Remove logs from the unsent metrics log as soon as they are made pending
This prevents a pending log from being duplicated if it is pending during shutdown (because of the code that saves the currently pending log to the unsent list). Since the write of the trimmed list back to the prefs system is still only done post-upload, it won't be lost during a crash.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/7248048