Revert "Linux: Remove some references to the sconsbuild directory."
This will fix dom_perf.py on Linux Perf (2).

extract_build.py is still extracting to sconsbuild,
so dom_perf.py gets confused and tries to run
performance_ui_tests in out/.  When it can't find it,
we see the classic error in our Buildbot instance of
a hanging bot and the rest of the steps aren't run.

The scope of this change is actually much bigger than
you may expect.  We have sconsbuild in many places,
and all of those places need to be updated with the
following logic: if 'out/Release' exists, use it,
otherwise use 'sconsbuild/Release'.  This is similar
to how runtest.py works (which is why runtest.py +
page cyclers + whatever elses use it didn't break with
this CL while the other tests did).

Once all of the places that determine the build dir on
Linux are updated with that logic, we then update the
master locations (eg. the memory master has paths that
refer to sconsbuild), land those changes, restart those
masters, then update extract_build.py to use a new path:
out instead of sconsbuild.

TBR=thestig@chromium.org
TBR=