Clean up .gypi and no-op GetCompositingSurface on TOUCH_UI.
GetCompositingSurface should just return a non-zero value on TOUCH_UI because we use a very different accelerated drawing path. In particular, the GPU process renders offscreen (to a region that it creates) and passes a handle to the video memory for the browser to display.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7314021