Remove DCHECK from HeartbeatSender::OnSignallingDisconnected()
Allow that method to be called without a corresponding call to
OnSignalConnected() - the linked bug-report explains how this can happen.

This is only a partial bug-fix; still need to address the problem of this
method being called on the wrong thread from ChromotingHost::Shutdown().

BUG=83788
TEST=Manual


Review URL: http://codereview.chromium.org/7307031