Fix extension docs generation
It turns out we were using a javascript reserved word as a bare identifier in
a jstemplate expression. We just need to prefix it with $this (bare identifiers
like this in jsdisplay expressions normally get an implicit $this).

BUG=88706
TEST=Run docs generation script at chrome/common/extensions/docs/build/build.py
and verify that all the "enum" spans don't lose their display: none style.

Review URL: http://codereview.chromium.org/7327028