Sync performance tests: simplifying timing procedure.
This change abandons the idea of disabling and enabling the network in order to measure commit time and update time separately.  Instead, we measure the time taken for a change by one client to propagate through the server to the second client.

When enabling and disabling the network, the timing measurements varied from ~2 to ~9 seconds and showed no trend.  With this change, we get less detailed results, but they match the expectation much more closely (~0.5 seconds for a single bookmark, up to ~5 seconds for 500).

I will continue to investigate this problem, but this change will allow me to obtain meaningful results as I move on to the other datatypes.

BUG=
TEST=


Review URL: http://codereview.chromium.org/7229019