Disable TCP connections for chromoting
Chromoting always runs PseudoTcp on top of jingle connections. Sometimes this
jingle connection is a TCP connection and it destroys performance completely.
Until we can bypass PseudoTcp when TCP connection is established and understand
the connection establishing algorithm we only UDP for now, which is known to
work best.

BUG=None
TEST=None

Review URL: http://codereview.chromium.org/7344002