Attempted fix for flakiness in RunExtensionTestIncognito.
My guess is that the problem is related to the the extension reloading when we
toggle incognito enabledness, and not waiting for the reload to finish.

Unfortunately I was unable to reproduce locally, so I'll have to see if this
helps on the buildbot.

BUG=88972
TEST=no


Review URL: http://codereview.chromium.org/7363012