NaCl: Enable tests on Windows
These tests have been working on the Windows NaCl/Chromium integration
bot for a while.

Set "--disable-tests" to disable a single test which currently fails in
the debug build.

Update DEPS to pull in the version of NaCl where the "--disable-tests"
option was added.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2039
TEST=nacl_integration step on the trybots

Review URL: http://codereview.chromium.org/7282035