PyAuto RPC API.
Adds a class, pyauto.RemoteProxy, that allows a test to establish a connection with a PyAuto instance on another machine. The other machine receives the connection by running the test remote_host.RemoteHost.RunHost. PyAuto calls can be made on the RemoteProxy object just as if it were a local PyUITest object.

BUG=chromium:89657
TEST=Manual. Not included. Basic instructions in pyauto.py.

Review URL: http://codereview.chromium.org/7285051