The problem:
When the user has the "Clear cookies and other site and plug-in data when I close my browser" setting enabled, all appcache data is deleted when the user closes the browser. The intention is to protect an appcache from deletion, if the user has installed a hosted app which refers to the origin of the appcache.

BUG=79604
TEST=ChromeAppCacheService.KeepOnDestruction, ChromeAppCacheService.RemoveOnDestruction


Review URL: http://codereview.chromium.org/7210006