Reland 93106 - Reparenting of RendererGLContexts supports going from no parent to having a parent.
RendererGLContext::SetParent is a no-op if the parent did not change to avoid a redundant sync IPC.

Added SetParentContext to WebGraphicsContext3DCommandBufferImpl. WebKit change to follow.

Original review: http://codereview.chromium.org/7237009

A patch that this was dependent on was reverting, causing this one to break 2D canvas. The patch is relanded. Relanding this one as well.
Review URL: http://codereview.chromium.org/7488001