During SetupSync, check for missing passphrase before waiting for sync
In SetupSync used by the sync integration tests, clients used to check
for a missing passphrase if initial sync failed, and return false if one
was required (indicating the fact that forward progress is impossible).

Due to a change in the sync startup flow, the check for a required
passphrase must now be done prior to waiting for initial sync.

This patch makes the above change, and thereby fixes the test case
SetPassphraseAndThenSetupSync that currently fails on the sync trybots.

TBR=zea@chromium.org
BUG=90371
TEST=sync_integration_tests

Review URL: http://codereview.chromium.org/7492044