AsyncHostResolver refactoring.
This change gets rid of the asymmetry between Resolve() calls that complete synchronously and those that complete asynchronously.  Makes code more readable/understandable, allows to embed OnStart/OnFinish/OnCancel calls within the Request::Request()/Request::On{Sync,Async}Complete/Request::~Request() so that we do not miss these inadvertently.  As a result of this, I found out that code logic as well as the unit test was not notifying the observer of the cancellation    due to object destruction.

BUG=60149
TEST=net_unittests --gtest_filter="AsyncHostResolverTest*"

Review URL: http://codereview.chromium.org/7492028