15 (fifteen) is the natural number following 14 and preceding 16. In English, it is the smallest natural number with seven letters in its spelled name. In speech, the numbers 15 and 50 are often confused. When carefully enunciated, they differ in which syllable is stressed: 15 /fɪfˈtiːn/ vs 50 /ˈfɪfti/. However, in dates such as 1500 ("fifteen hundred") or when contrasting numbers in the teens, the stress generally shifts to the first syllable: 15 /ˈfɪftiːn/.
Fifteen is:
The number of days in each of the 24 cycles of the Chinese calendar.
The number of guns in a gun salute to Army, Marine Corps, and Air Force Lieutenant Generals, and Navy and Coast Guard Vice Admirals.
The designation of Interstate 15, a freeway that runs from California to Montana.
The designation of U.S. Route 15, a highway that runs from South Carolina to New York.
The number of checkers each side has at the start of a backgammon game.
The 15 puzzle.
A restaurant in London, built by celebrity chef Jamie Oliver
15 minutes of fame
The 15th is a Wire song
The number of balls in the eight ball variant of billiards
The number of function keys on most Mac keyboards
The number of letters in the words "uncopyrightable", "dermatoglyphics", "misconjugatedly", and "hydropneumatics", which are the longest words in the English language that do not repeat a letter.[citation needed]
15 is one of The Numbers - 4, 8, 15, 16, 23, and 42 - featured in Lost.
The number to dial for SAMU in case of an emergency in France.
The number of minutes in one quarter of an hour; 15 minutes past or before an hour is often known as quarter past and quarter to, respectively.
The number of the French department Cantal
The number corresponding to The Devil in tarot cards.

Credit: http://en.wikipedia.org/wiki/15_(number)