Merge 94152 - Ensures the throbber is turned off when more than one page of history is available.
The throbber is now automatically turned on when getting or searching history. However, it was not turned back off in the event that more than one page of history is available.

This bug was introduced in http://codereview.chromium.org/7399007

BUG=90388
TEST=Ensure you have more than 150 items in your history. Hit CTRL-H. After displaying the history the throbber should stop and you should get an "Older" link at the bottom.
Review URL: http://codereview.chromium.org/7484071

TBR=beaudoin@chromium.org
Review URL: http://codereview.chromium.org/7517014