Fix method call on potentially deleted object in NativeWidgetViews.
The problem was that I was unconditionally calling set_delete_native_widget() from the NativeWidgetViews destructor, which is wrong if the NativeWidgetView has already been destroyed with its view hierarchy.  The unconditional call is correct in the NativeWidgetView destructor because it must prevent itself from being deleted again even if it does not delete the NativeWidgetViews (WIDGET_OWNS_NATIVE_WIDGET case) and is safe because the NativeWidgetViews lifetime is always longer than its NativeWidgetView.

BUG=chromium:90484
TEST=valgrind out/Debug/views_unittests --gtest_filter="WidgetOwnershipTest.*ViewsNativeWidget*" --single-process


Review URL: http://codereview.chromium.org/7461082