Re-enable pre- and auto-login after the M14 branch is made.
BUG=87151
TEST=There should be no options in about:flags for either pre-login or
auto-login.  There is now a checkbox in the "Passwords" section of the 
"Personal stuff" tab in options page called "Enable automatic sign in for Google
sites", which is on by default in new profiles.  When checked, both pre- and
auto-login are enabled.  When unchechecked, both are disabled.
The user can also disable automatic sign in using the "Never offer to sign me
in" button in the auto-login infobar.

Review URL: http://codereview.chromium.org/7470019