mac: include .so files in framework copy
The -C flag to rsync obeys CVS ignore rules, which means we would
drop .so files when copying files to "unversion" a framework.

This change adjusts the rsync flags to explicitly include .so files.

This is necessary to support a separate change, which renames
a .dylib to a .so.

Review URL: http://codereview.chromium.org/7519011