Merge 94382 - [Sync] Fix sync UI handling of passphrases.
The UI's handling of which passphrase to use now depends on which UI is visible and is handled a bit more elegantly. We now send down an empty passphrase if that's what the user entered. If encrypted datatypes are enabled, this will trigger a passphrase failure, prompting for a new passphrase. If they are not enabled, the passphrase is silently accepted, until the user attempts to enable an encrypted type.

BUG=90603
TEST=Sign into sync with an account with existing passphrase. Enter empty passphrase.

Review URL: http://codereview.chromium.org/7514032

TBR=zea@chromium.org
Review URL: http://codereview.chromium.org/7519021