Merge trunk r94279 to the 14.0.835 branch.
[Mac] Make Chrome not delay system sleep

Some processes [e.g. the GPU process] do not set up a SystemMonitor object, since the startup code was changed to always allocate the Power notification IO port [http://codereview.chromium.org/7235023/] these processes were causing a delay in sleep as the power manager was trying to contact the process.

This change whitelists the IO port allocation in ChromeMain() to only include process types that allocate a SystemMonitor which fixes the sleep issue.

If a process not in the whitelist attempts to use the SystemMonitor it will hit a DCHECK in a debug build and the error will be ignored in release (see SystemMonitor::PlatformInit()).

BUG=88867
TEST=When Chrome is running, Macs should sleep immediately.

Review URL: http://codereview.chromium.org/7491038

TBR=jeremy@chromium.org
Review URL: http://codereview.chromium.org/7492061