Adjust exception handling syntax in sync test server.
The newly added sync perf tests fail on mac 10.5, but succeed on all other platforms.
On mac 10.5, the sync test server cannot be launched due to a syntax error.  It turns
out that the sync test server had two instances of the except-as syntax, which was
added in Python3 and ported to 2.6, but not 2.5, which is the default on mac 10.5.

BUG=91006
TEST=sync_performance_tests, sync_integration_tests

Review URL: http://codereview.chromium.org/7531031