Re-enable TaskManagerBrowserTest.ReloadExtension
In the past this test had been crashy, and I was able to reproduce the crash
occasionally in vmware and on the try bots. I'm no longer able to repro, so
I'm speculatively re-enabling it to see if the underlying cause is actually
fixed.

BUG=42315
TEST=none

TBR=mpcomplete@chromium.org

Review URL: http://codereview.chromium.org/7537006