history: use display title, not page title
entry.title() is the underlying page title, but we want to record
in history the actual title we show to the user.  These differ in
corner cases like when the page lacks a title, and also importantly
in that the display title can include RTL metadata.

BUG=27094

Review URL: http://codereview.chromium.org/7453053