Fix the issue of virtual keyobard overlapping input method candidates area.
This CL fixes most cases of the issue, but not completely.
When input method produces many candidates with many characters, right part of candidate run off the edge of the screen. In order to handle such cases, we need to show candidates more cleverly.

BUG=none
TEST=Manually on the machine.

Review URL: http://codereview.chromium.org/7564004