Turn on --enable-nacl-debug.
Reenable the --enable-nacl-debug switch.

This was accidentally turned off in a bigger CL:
  http://codereview.chromium.org/7395024/
See nacl_main.cc and nacl_listener.cc in that CL to compare to this.

BUG= http://code.google.com/p/chromium/issues/detail?id=91425

TEST= manual - w/o this fix, chrome --enable-nacl-debug no longer works, with
this change, I was able to launch Chrome from Visual Studio and set breakpoints
in a nexe.
Review URL: http://codereview.chromium.org/7547002