Allow sync integration tests to operate on multiple datatypes: Autofill
The sync integration tests currently use a class hierarchy where the test classes for each datatype are subclasses of LiveSyncTest. While this design worked in the past, it allows tests to work with only one datatype at a time, and therefore doesn't allow us to test the interplay between datatypes.

This patch is another in the series of patches that will move away from an inheritance model to one where test cases can operate on more than one datatype. It updates the Autofill datatype to the new model, and contains the following changes:

- LiveAutofillSyncTest no longer inherits from LiveSyncTest, but is renamed to AutofillHelper, and contains a bunch of static methods that perform various operations related to Autofill.
- TwoClientLiveAutofillSyncTest is renamed to TwoClientAutofillSyncTest, and is a subclass of LiveSyncTest. It uses the methods in AutofillHelper by including its header file and attaching itself to the helper.

BUG=88510
TEST=sync_integration_tests, sync_performance_tests

Review URL: http://codereview.chromium.org/7536001