Profiles: Introduce AllBrowserContextsAndSources().
Eventually, we need to remove all current instances of AllSources(). One by
one, they need to either changed to listen to a specific notification source,
or (in very few cases where we actually are listening for a global
notification) they need to work correctly in the presence of multiple profiles,
in which case they'll be turned to AllBrowserContextsAndSources().

This patch, in addition:
- Deletes more unused notifications.
- Changes a few AllSources() calls to listen to specific objects.

BUG=87457
TEST=none


Review URL: http://codereview.chromium.org/7548022