Set send buffer size for the raw socket used by chromoting
On Windows the send buffer size of UDP socket has a much smaller value
than the PseudoTcp's send buffer size. This causes PseudoTcp to flood
the UDP socket and results in a EWOULDBLOCK, however the code doesn't
handle this properly but simply treat the packet as lost. This happens
very often and causes Windows host to be very janky.

BUG=91495
TEST=None

Review URL: http://codereview.chromium.org/7508015